ai-multiten-full/backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
motor>=3.1
pydantic>=2.0
python-jose>=3.3.0
python-dotenv>=1.0
typing-extensions
